+++ packages/ffe-account-selector-react/src/subcomponents/suggestion/SuggestionListContainer.js
@@ -9,7 +9,9 @@ import SuggestionList from './SuggestionList';
 class SuggestionListContainer extends React.Component {
     constructor(props) {
         super(props);
-        this.refHighlightedSuggestion = this.refHighlightedSuggestion.bind(this);
+        this.refHighlightedSuggestion = this.refHighlightedSuggestion.bind(
+            this,
+        );
     }
 
     refHighlightedSuggestion(suggestionEl) {
@@ -35,7 +37,10 @@ class SuggestionListContainer extends React.Component {
 
     setScrollPosPrevious() {
         const { highlightedIndex } = this.props;
-        this._setScrollPos(highlightedIndex * this.highlightedSuggestionHeight - this.highlightedSuggestionHeight);
+        this._setScrollPos(
+            highlightedIndex * this.highlightedSuggestionHeight -
+                this.highlightedSuggestionHeight,
+        );
     }
 
     render() {