+++ packages/ffe-checkbox-react/docs/example-component.js
@@ -2,11 +2,7 @@ import React from 'react';
 import Checkbox from '../src/Checkbox';
 
 export const MainExample = (
-    <Checkbox
-        checked={false}
-        name="Yellow"
-        label="Yellow"
-    />
+    <Checkbox checked={false} name="Yellow" label="Yellow" />
 );
 
 export default props => <Checkbox {...props} />;