+++ packages/ffe-form-react/src/Input.js
@@ -1,25 +1,17 @@
 import React from 'react';
-import {
-    bool,
-    string
-} from 'prop-types';
+import { bool, string } from 'prop-types';
 import classNames from 'classnames';
 
-const Input = ({
-    className,
-    inline,
-    textLike,
-    ...rest
-}) => {
+const Input = ({ className, inline, textLike, ...rest }) => {
     return (
         <input
-            className={ classNames(
+            className={classNames(
                 'ffe-input-field',
                 { 'ffe-input-field--inline': inline },
                 { 'ffe-input-field--text-like': textLike },
-                className
-            ) }
-            { ...rest }
+                className,
+            )}
+            {...rest}
         />
     );
 };
@@ -29,7 +21,7 @@ Input.propTypes = {
     /** Input fields default to `display: block;`. Set this to `true` to apply the inline modifier. */
     inline: bool,
     /** Apply the text-like modifier by setting this to `true`. */
-    textLike: bool
+    textLike: bool,
 };
 
 export default Input;