+++ packages/ffe-form-react/src/SuccessFieldMessage.js
@@ -3,12 +3,7 @@ import { node } from 'prop-types';
 import FieldMessage from './FieldMessage';
 
 const SuccessFieldMessage = props => {
-    return (
-        <FieldMessage
-            { ...props }
-            type="success"
-        />
-    );
+    return <FieldMessage {...props} type="success" />;
 };
 
 export default SuccessFieldMessage;