+++ packages/ffe-radio-button-react/src/radio-base.test.js
@@ -10,7 +10,9 @@ import RadioBase from './radio-base';
 describe('<RadioBase />', () => {
     describe('rendering', () => {
         describe('default', () => {
-            const wrapper = shallow(<RadioBase labelClasses="" value="shark" label="Kjempetorsk" />);
+            const wrapper = shallow(
+                <RadioBase labelClasses="" value="shark" label="Kjempetorsk" />,
+            );
 
             it('should render an input with type radio', () => {
                 expect(wrapper)
@@ -49,15 +51,24 @@ describe('<RadioBase />', () => {
         });
 
         it('should apply labelClasses to the label', () => {
-            const wrapper = shallow(<RadioBase labelClasses="test-class" value="" />);
+            const wrapper = shallow(
+                <RadioBase labelClasses="test-class" value="" />,
+            );
             expect(wrapper.find('label').hasClass('test-class')).to.be.true;
         });
 
         it('should apply pass-through extra props', () => {
             const wrapper = shallow(
-                <RadioBase labelClasses="" value="shark" label="Kjempetorsk" aria-describedby="svorsk" />,
+                <RadioBase
+                    labelClasses=""
+                    value="shark"
+                    label="Kjempetorsk"
+                    aria-describedby="svorsk"
+                />,
             );
-            expect(wrapper.find('input').prop('aria-describedby')).to.equal('svorsk');
+            expect(wrapper.find('input').prop('aria-describedby')).to.equal(
+                'svorsk',
+            );
         });
     });
 
@@ -70,7 +81,9 @@ describe('<RadioBase />', () => {
 
         it('should call onChange when clicked', () => {
             const spy = sinon.spy();
-            const wrapper = shallow(<RadioBase onChange={spy} labelClasses="" />);
+            const wrapper = shallow(
+                <RadioBase onChange={spy} labelClasses="" />,
+            );
             wrapper.find('input').simulate('change');
             expect(spy.calledOnce).to.be.true;
         });
@@ -79,8 +92,18 @@ describe('<RadioBase />', () => {
             const spy = sinon.spy();
             const wrapper = mount(
                 <fieldset>
-                    <RadioBase name="fish" value="hai" onChange={spy} labelClasses="" />
-                    <RadioBase name="fish" value="kjempetorsk" onChange={spy} labelClasses="" />
+                    <RadioBase
+                        name="fish"
+                        value="hai"
+                        onChange={spy}
+                        labelClasses=""
+                    />
+                    <RadioBase
+                        name="fish"
+                        value="kjempetorsk"
+                        onChange={spy}
+                        labelClasses=""
+                    />
                 </fieldset>,
             );
             const shark = wrapper.find('input[value="hai"]');
@@ -102,64 +125,108 @@ describe('<RadioBase />', () => {
         it('sets missing prop to false', () => {
             const wrapper = shallow(<RadioBase value="Hei" labelClasses="" />);
 
-            expect(wrapper.find('input').prop('aria-invalid')).to.equal('false');
-            expect(wrapper.find('label').prop('className')).not.to.equal('ffe-radio-button--invalid');
+            expect(wrapper.find('input').prop('aria-invalid')).to.equal(
+                'false',
+            );
+            expect(wrapper.find('label').prop('className')).not.to.equal(
+                'ffe-radio-button--invalid',
+            );
         });
 
         it('handles boolean `false` aria-invalid-prop', () => {
-            const wrapper = shallow(<RadioBase value="Hei" labelClasses="" aria-invalid={false} />);
+            const wrapper = shallow(
+                <RadioBase value="Hei" labelClasses="" aria-invalid={false} />,
+            );
 
-            expect(wrapper.find('input').prop('aria-invalid')).to.equal('false');
-            expect(wrapper.find('label').prop('className')).not.to.equal('ffe-radio-button--invalid');
+            expect(wrapper.find('input').prop('aria-invalid')).to.equal(
+                'false',
+            );
+            expect(wrapper.find('label').prop('className')).not.to.equal(
+                'ffe-radio-button--invalid',
+            );
         });
 
         it('handles boolean `true` aria-invalid-prop', () => {
-            const wrapper = shallow(<RadioBase value="Hei" labelClasses="" aria-invalid={true} />);
+            const wrapper = shallow(
+                <RadioBase value="Hei" labelClasses="" aria-invalid={true} />,
+            );
 
             expect(wrapper.find('input').prop('aria-invalid')).to.equal('true');
-            expect(wrapper.find('label').prop('className')).to.equal('ffe-radio-button--invalid');
+            expect(wrapper.find('label').prop('className')).to.equal(
+                'ffe-radio-button--invalid',
+            );
         });
 
         it('handles string "false" aria-invalid-prop', () => {
-            const wrapper = shallow(<RadioBase value="Hei" labelClasses="" aria-invalid="false" />);
+            const wrapper = shallow(
+                <RadioBase value="Hei" labelClasses="" aria-invalid="false" />,
+            );
 
-            expect(wrapper.find('input').prop('aria-invalid')).to.equal('false');
-            expect(wrapper.find('label').prop('className')).not.to.equal('ffe-radio-button--invalid');
+            expect(wrapper.find('input').prop('aria-invalid')).to.equal(
+                'false',
+            );
+            expect(wrapper.find('label').prop('className')).not.to.equal(
+                'ffe-radio-button--invalid',
+            );
         });
 
         it('handles string "true" aria-invalid-prop', () => {
-            const wrapper = shallow(<RadioBase value="Hei" labelClasses="" aria-invalid="true" />);
+            const wrapper = shallow(
+                <RadioBase value="Hei" labelClasses="" aria-invalid="true" />,
+            );
 
             expect(wrapper.find('input').prop('aria-invalid')).to.equal('true');
-            expect(wrapper.find('label').prop('className')).to.equal('ffe-radio-button--invalid');
+            expect(wrapper.find('label').prop('className')).to.equal(
+                'ffe-radio-button--invalid',
+            );
         });
 
         it('handles boolean `false` invalid-prop', () => {
-            const wrapper = shallow(<RadioBase value="Hei" labelClasses="" invalid={false} />);
+            const wrapper = shallow(
+                <RadioBase value="Hei" labelClasses="" invalid={false} />,
+            );
 
-            expect(wrapper.find('input').prop('aria-invalid')).to.equal('false');
-            expect(wrapper.find('label').prop('className')).not.to.equal('ffe-radio-button--invalid');
+            expect(wrapper.find('input').prop('aria-invalid')).to.equal(
+                'false',
+            );
+            expect(wrapper.find('label').prop('className')).not.to.equal(
+                'ffe-radio-button--invalid',
+            );
         });
 
         it('handles boolean `true` invalid-prop', () => {
-            const wrapper = shallow(<RadioBase value="Hei" labelClasses="" invalid={true} />);
+            const wrapper = shallow(
+                <RadioBase value="Hei" labelClasses="" invalid={true} />,
+            );
 
             expect(wrapper.find('input').prop('aria-invalid')).to.equal('true');
-            expect(wrapper.find('label').prop('className')).to.equal('ffe-radio-button--invalid');
+            expect(wrapper.find('label').prop('className')).to.equal(
+                'ffe-radio-button--invalid',
+            );
         });
 
         it('handles string "false" invalid-prop', () => {
-            const wrapper = shallow(<RadioBase value="Hei" labelClasses="" invalid="false" />);
+            const wrapper = shallow(
+                <RadioBase value="Hei" labelClasses="" invalid="false" />,
+            );
 
-            expect(wrapper.find('input').prop('aria-invalid')).to.equal('false');
-            expect(wrapper.find('label').prop('className')).not.to.equal('ffe-radio-button--invalid');
+            expect(wrapper.find('input').prop('aria-invalid')).to.equal(
+                'false',
+            );
+            expect(wrapper.find('label').prop('className')).not.to.equal(
+                'ffe-radio-button--invalid',
+            );
         });
 
         it('handles string "true" invalid-prop', () => {
-            const wrapper = shallow(<RadioBase value="Hei" labelClasses="" invalid="true" />);
+            const wrapper = shallow(
+                <RadioBase value="Hei" labelClasses="" invalid="true" />,
+            );
 
             expect(wrapper.find('input').prop('aria-invalid')).to.equal('true');
-            expect(wrapper.find('label').prop('className')).to.equal('ffe-radio-button--invalid');
+            expect(wrapper.find('label').prop('className')).to.equal(
+                'ffe-radio-button--invalid',
+            );
         });
     });
 });