+++ styleguide.config.js
@@ -32,6 +32,10 @@ module.exports = {
                 'packages/ffe-core-react/src/typography/[A-Z]+([A-Za-z]).js',
         },
         {
+            name: 'Ikoner',
+            content: 'packages/ffe-icons-react/USAGE.md',
+        },
+        {
             name: 'Kakediagram',
             content: 'packages/ffe-chart-donut-react/USAGE.md',
             components:
@@ -64,6 +68,11 @@ module.exports = {
             name: 'Grid',
             content: 'packages/ffe-grid-react/USAGE.md',
             components: 'packages/ffe-grid-react/src/[A-Z]+([A-Za-z]).js',
+        },
+        {
+            name: 'Tabeller',
+            content: 'packages/ffe-tables-react/USAGE.md',
+            components: 'packages/ffe-tables-react/src/Table.js',
         },
         {
             name: 'Skjema',