+++ packages/ffe-context-message-react/CHANGELOG.md
@@ -3,13 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-      <a name="3.0.4"></a>
+    <a name="3.0.6"></a>
+
+## [3.0.6](***REMOVED***) (2018-01-10)
+
+**Note:** Version bump only for package ffe-context-message-react
+
+    <a name="3.0.5"></a>
+
+## [3.0.5](***REMOVED***) (2018-01-10)
+
+**Note:** Version bump only for package ffe-context-message-react
+
+<a name="3.0.4"></a>
 
 ## [3.0.4](***REMOVED***) (2018-01-09)
 
 **Note:** Version bump only for package ffe-context-message-react
 
-      <a name="3.0.3"></a>
+<a name="3.0.3"></a>
 
 ## [3.0.3](***REMOVED***) (2018-01-05)